components/agent.py
from langchain import OpenAI, LLMChain
from langchain_community.tools.tavily_search import TavilySearchResults
from langchain_core.messages import HumanMessage


from langchain.prompts import PromptTemplate
from langchain.agents import initialize_agent, Tool

